Removing event handling result
