+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/event/handler/chat/DefaultChatEventHandler.kt
@@ -18,9 +18,7 @@ package io.getstream.chat.android.offline.event.handler.chat
 
 import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.events.ChatEvent
-import io.getstream.chat.android.client.events.ChatEventHandler
 import io.getstream.chat.android.client.events.CidEvent
-import io.getstream.chat.android.client.events.EventHandlingResult
 import io.getstream.chat.android.client.events.HasChannel
 import io.getstream.chat.android.client.events.MemberAddedEvent
 import io.getstream.chat.android.client.events.MemberRemovedEvent
