+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/state/querychannels/QueryChannelsState.kt
@@ -19,8 +19,8 @@ package io.getstream.chat.android.offline.plugin.state.querychannels
 import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
 import io.getstream.chat.android.client.api.models.querysort.QuerySorter
-import io.getstream.chat.android.client.events.ChatEventHandler
 import io.getstream.chat.android.client.models.Channel
+import io.getstream.chat.android.offline.event.handler.chat.ChatEventHandler
 import io.getstream.chat.android.offline.event.handler.chat.factory.ChatEventHandlerFactory
 import kotlinx.coroutines.flow.StateFlow
